+++ convex/auth.config.ts
@@ -1,4 +1,4 @@
-export default {
+const authConfig = {
   providers: [
     {
       domain: "https://securetoken.google.com/reindex-94dbc",
@@ -6,3 +6,5 @@ export default {
     },
   ],
 };
+
+export default authConfig;